Avoid core changes and move files from philu_overrides to openedx/features
