+++ openedx/core/djangoapps/user_api/legacy_urls.py
@@ -8,6 +8,7 @@ from rest_framework import routers
 
 from . import views as user_api_views
 from lms.djangoapps.philu_overrides.user_api import views as user_api_views_custom
+from openedx.features.student_account import views as student_account_views_custom
 from .models import UserPreference
 
 USER_API_ROUTER = routers.DefaultRouter()
@@ -40,7 +41,7 @@ if settings.FEATURES.get('ENABLE_COMBINED_LOGIN_REGISTRATION'):
     urlpatterns += [
         url(r'^v1/account/login_session/$', user_api_views_custom.LoginSessionViewCustom.as_view(),
             name="user_api_login_session"),
-        url(r'^v1/account/registration/$', user_api_views_custom.RegistrationViewCustom.as_view(),
+        url(r'^v1/account/registration/$', student_account_views_custom.RegistrationViewCustom.as_view(),
             name="user_api_registration"),
         url(r'^v1/account/password_reset/$', user_api_views.PasswordResetView.as_view(),
             name="user_api_password_reset"),
